requirements/dev.txt
fastapi
sqlalchemy
uvicorn[standard]
psycopg2-binary
alembic
pydantic[email]
python-dotenv